move to fluxie dir
