+++ fluxie/operators/rolling_mean.py
